💄style : 캘린더 파일의 스타일 컴포넌트를 분리합니다.
